actions: Massive cleanup
We now use a better IniFile class which is closer to the metal
(RawConfigParser subclass). The downside is we have some increased
complexity when merging files as we implement list values and those need
to be merged properly.

Additionally, the entire actions module has been rewritten to be
agnostic in name to the MimeType class. This is to prepare for
Category/Intent/etc support in the actions files.